+++ GrandCentralDispatch/Nodes/Remote/Unary/UnarySequentialDispatcherRemoteNode.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reactive.Concurrency;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
@@ -21,7 +20,7 @@ using GrandCentralDispatch.Remote;
 namespace GrandCentralDispatch.Nodes.Remote.Unary
 {
     /// <summary>
-    /// Node which process items sequentially.
+    /// Node which process items sequentially remotely.
     /// </summary>
     /// <typeparam name="TInput">Item to be processed</typeparam>
     internal sealed class UnarySequentialDispatcherRemoteNode<TInput> : UnarySequentialProcessor<TInput>,
@@ -140,7 +139,7 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
                 {
                     NodeMetrics.RemoteNodeHealth = remoteNodeHealth;
                 });
-            _nodeHub = StreamingHubClient.Connect<INodeHub, INodeReceiver>(channel, (INodeReceiver)nodeReceiver);
+            _nodeHub = StreamingHubClient.Connect<INodeHub, INodeReceiver>(channel, (INodeReceiver) nodeReceiver);
 
             NodeMetrics = new NodeMetrics(Guid.NewGuid());
         }
@@ -179,7 +178,7 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
                         if (CpuUsage > _clusterOptions.LimitCpuUsage)
                         {
                             var suspensionTime = (CpuUsage - _clusterOptions.LimitCpuUsage) / CpuUsage * 100;
-                            await Task.Delay((int)suspensionTime, ct);
+                            await Task.Delay((int) suspensionTime, ct);
                         }
 
                         var persistentCacheToken = new CancellationTokenSource();
@@ -200,7 +199,7 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
                     {
                         Interlocked.Increment(ref _processedItems);
                         currentProgress++;
-                        progress.Report(currentProgress / bulk.Count);
+                        progress.Report((double) currentProgress / bulk.Count);
                     }
                 }, cancellationToken).ConfigureAwait(false);
 
@@ -248,7 +247,7 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
                         if (CpuUsage > _clusterOptions.LimitCpuUsage)
                         {
                             var suspensionTime = (CpuUsage - _clusterOptions.LimitCpuUsage) / CpuUsage * 100;
-                            await Task.Delay((int)suspensionTime, ct);
+                            await Task.Delay((int) suspensionTime, ct);
                         }
 
                         var entity = item();
@@ -270,7 +269,7 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
                     {
                         Interlocked.Increment(ref _executorProcessedItems);
                         currentProgress++;
-                        progress.Report(currentProgress / bulk.Count);
+                        progress.Report((double) currentProgress / bulk.Count);
                     }
                 }, cancellationToken).ConfigureAwait(false);
 
@@ -316,7 +315,8 @@ namespace GrandCentralDispatch.Nodes.Remote.Unary
             if (NodeMetrics == null) return;
             NodeMetrics.TotalItemsProcessed = TotalItemsProcessed();
             NodeMetrics.ItemsEvicted = ItemsEvicted();
-            NodeMetrics.CurrentThroughput = Interlocked.Exchange(ref _processedItems, 0L) + Interlocked.Exchange(ref _executorProcessedItems, 0L);
+            NodeMetrics.CurrentThroughput = Interlocked.Exchange(ref _processedItems, 0L) +
+                                            Interlocked.Exchange(ref _executorProcessedItems, 0L);
             NodeMetrics.BufferSize = GetBufferSize();
             NodeMetrics.Full = IsFull();
             try
